requirements.txt
certifi==2019.11.28
numpy==1.18.2
opencv-python==4.1.1.26
Pillow==7.0.0
